SIV-638 A typo in package name fixed and wildcard removed
