+++ src/main/java/uk/gov/companieshouse/officer/delta/processor/transformer/TransformerUtils.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.officer.delta.processor.tranformer;
+package uk.gov.companieshouse.officer.delta.processor.transformer;
 
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
