+++ src/test/java/uk/gov/companieshouse/officer/delta/processor/transformer/PrincipalOfficeAddressTransformTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.officer.delta.processor.tranformer;
+package uk.gov.companieshouse.officer.delta.processor.transformer;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
