+++ src/test/java/uk/gov/companieshouse/officer/delta/processor/transformer/TransformerUtilsTest.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.officer.delta.processor.tranformer;
+package uk.gov.companieshouse.officer.delta.processor.transformer;
 
 
 import static java.time.ZoneOffset.UTC;
